Reinforcement learning
Neural network that plays Super Mario. The goal is to train a neural network known as Proximal Policy Optimization (PPO) to learn how to play Super Mario. The python package used is stable baseline and can be found on this webpage https://stable-baselines.readthedocs.io/en/master/modules/ppo2.html . This neural network was trained for a million iterations and the first level was passed. However, if you want to pass the next level you should train for more times.